split out category code into its own resource
