+++ server/src/api/forum/categories/api.go
@@ -5,17 +5,17 @@ import (
 	"go.uber.org/fx"
 
 	"github.com/openmultiplayer/web/server/src/authentication"
-	"github.com/openmultiplayer/web/server/src/resources/forum"
+	"github.com/openmultiplayer/web/server/src/resources/forum/category"
 )
 
 type service struct {
 	as   *authentication.State
-	repo forum.Repository
+	repo category.Repository
 }
 
 func Build() fx.Option {
 	return fx.Options(
-		fx.Provide(func(as *authentication.State, repo forum.Repository) *service { return &service{as, repo} }),
+		fx.Provide(func(as *authentication.State, repo category.Repository) *service { return &service{as, repo} }),
 		fx.Invoke(func(
 			r chi.Router,
 			s *service,
