+++ server/src/bsworker/bsworker.go
@@ -8,13 +8,15 @@ import (
 
 	"github.com/openmultiplayer/web/server/src/resources/burgershot"
 	"github.com/openmultiplayer/web/server/src/resources/forum"
+	"github.com/openmultiplayer/web/server/src/resources/forum/category"
 	"github.com/openmultiplayer/web/server/src/resources/user"
 )
 
 type Worker struct {
-	bs    burgershot.Repository
-	users user.Repository
-	forum forum.Repository
+	bs       burgershot.Repository
+	users    user.Repository
+	forum    forum.Repository
+	category category.Repository
 }
 
 func New(
@@ -22,9 +24,10 @@ func New(
 	bs burgershot.Repository,
 	users user.Repository,
 	forum forum.Repository,
+	category category.Repository,
 ) Worker {
 
-	w := Worker{bs, users, forum}
+	w := Worker{bs, users, forum, category}
 
 	wctx, stop := context.WithCancel(context.Background())
 	lc.Append(fx.Hook{
